feat: API 통신을 위한 응답 규격 세팅 #16
